Improvements to CLI

+++ markingpy/cli.py
@@ -1,6 +1,7 @@
 """Command line interface for MarkingPy."""
 
 import sys
+import statistics
 from argparse import ArgumentParser
 from pathlib import Path
 
@@ -18,51 +19,87 @@ def run():
     General markingpy cli
     """
     config = GLOBAL_CONF
-    parser = ArgumentParser()
+    parser = ArgumentParser(
+        description=(
+            "Markingpy grading tool."
+        )
+    )
     args = parser.parse_args()
     return 0
 
 
-def is_markscheme(path):
-    is_ms = False
-    if not isinstance(path, Path):
-        return is_ms
-
-    if not path.name.endswith(".py"):
-        return is_ms
-
-    if not path.exists():
-        return is_ms
-
-    text = path.read_text()
-    if "import markingpy" in text or "from markingpy import" in text:
-        is_ms = True
-    return is_ms
 
 
 class MarkschemeCommands:
 
     @staticmethod
     def run(markscheme, cli_args):
-        parser = ArgumentParser()
-        parser.add_argument("--style-formula", type=str)
-        parser.add_argument("--style-marks", type=int)
+        parser = ArgumentParser(
+            usage=f'markingpy {sys.argv[1]} run [OPTIONS]',
+            description=(
+                "Run the markingpy grader using the marking scheme "
+                f"{sys.argv[1]}. The grader will use the settings "
+                f"defined in {sys.argv[1]} along with any default "
+                "values (set in .markingpy if this exists) or options passed "
+                "to this command."
+            )
+        )
+
+        parser.add_argument(
+            "--style-formula",
+            type=str,
+            help="Formula used to generate a score from the linter report."
+        )
+        parser.add_argument(
+            "--style-marks",
+            type=int,
+            help="Number of marks available for coding style."
+        )
         parser.add_argument(
             "--score-style",
-            choices=["real", "percentage", "marks/total", "all"],
-            default="all",
+            help="Formatting style for marks to be displayed in feedback."
         )
-        parser.add_argument("--submission-path", type=str)
-        parser.add_argument("--marks-db", type=str)
+        parser.add_argument(
+            "--submission-path",
+            type=str,
+            help="Path to submissions."
+        )
+        parser.add_argument(
+            "--marks-db",
+            type=str,
+            help="Path to database to store submission results and feedback."
+        )
+        parser.add_argument(
+            "--marks",
+            type=int,
+            help=(
+                "Total marks available for this coursework." 
+                " This option is only used for validation."
+            )
+        )
         markscheme.update_config(vars(parser.parse_args(cli_args)))
+        markscheme.validate()
         grader = Grader(markscheme)
         with grader:
             grader.grade_submissions()
 
     @staticmethod
     def grades(markscheme, cli_args):
-        parser = ArgumentParser()
-        parser.add_argument("--marks-db", type=str)
+        parser = ArgumentParser(
+            usage=f'markingpy {sys.argv[1]} grades [OPTIONS]',
+            description=(
+                "Print the grades obtained in the last run of the grader"
+                "with this marking scheme. This will load the results in"
+                "the default database, the database specified in "
+                f"{sys.argv[1]} or the database specified in the options for "
+                "this command."
+            )
+        )
+        parser.add_argument(
+            "--marks-db",
+            type=str,
+            help="Path to database to store submission results and feedback."
+        )
         markscheme.update_config(vars(parser.parse_args(cli_args)))
         subs = markscheme.get_db().fetch_all()
         for sid, perc, _ in subs:
@@ -70,10 +107,21 @@ class MarkschemeCommands:
 
     @staticmethod
     def summary(markscheme, cli_args):
-        import statistics
-
-        parser = ArgumentParser()
-        parser.add_argument("--marks-db", type=str)
+        parser = ArgumentParser(
+            usage=f'markingpy {sys.argv[1]} summary [OPTIONS]',
+            description=(
+                "Print a summary of grades obtained in the last run of the "
+                "grader with this marking scheme. This will load the results "
+                "in the default database, the database specified in "
+                f"{sys.argv[1]} or the database specified in the options for "
+                "this command."
+            )
+        )
+        parser.add_argument(
+            "--marks-db",
+            type=str,
+            help="Path to database to store submission results and feedback."
+        )
         markscheme.update_config(vars(parser.parse_args(cli_args)))
         subs = markscheme.get_db().fetch_all()
         percs = [sub[1] for sub in subs]
@@ -86,11 +134,33 @@ class MarkschemeCommands:
 
     @staticmethod
     def dump(markscheme, cli_args):
-        parser = ArgumentParser()
-        parser.add_argument("--marks-db", type=str)
-        parser.add_argument("path", default=".", nargs="?")
+        parser = ArgumentParser(
+            usage=f'markingpy {sys.argv[1]} dump [OPTIONS] path',
+            description=(
+                "Dump the feedback obtained in the last run of the "
+                "grader with this marking scheme into the directory 'path'. "
+                "This will load the results in the default database, "
+                f"the database specified in {sys.argv[1]} "
+                "or the database specified in the options for "
+                "this command."
+            )
+        )
+        parser.add_argument(
+            "--marks-db",
+            type=str,
+            help="Path to database to store submission results and feedback."
+        )
+        parser.add_argument(
+            "path",
+            default=".",
+            nargs="?",
+            help="Directory to populate with feedback."
+        )
         args = vars(parser.parse_args(cli_args))
         path = Path(args.pop("path"))
+        if path.exists() and not path.is_dir():
+            raise ValueError('{path} is not a directory')
+        path.mkdir(exists_ok=True)
         markscheme.update_config(args)
         for sub_id, _, fb in markscheme.get_db().fetch_all():
             (path / (sub_id + ".txt")).write_text(fb)
@@ -115,8 +185,7 @@ def main():
     except IndexError:
         cmd = "run"
         args = []
-    if not is_markscheme(path):
-        return run()
+
 
     markscheme = import_markscheme(path)
     try:
